+++ Sources/SwiftyHYGDB/Star3D+CVS.swift
@@ -38,7 +38,7 @@ extension Star3D {
             let dist: Double = readNumber(at: &index, stringPtr: rowPtr) else { return nil }
         let pmra: Double? = advanceByYears != nil ? readNumber(at: &index, stringPtr: rowPtr) : nil
         let pmdec: Double? = advanceByYears != nil ? readNumber(at: &index, stringPtr: rowPtr) : nil
-        let rv: Double? = readNumber(at: &index, stringPtr: rowPtr)
+        let rv: Float? = readNumber(at: &index, stringPtr: rowPtr)
         guard let mag: Float = readNumber(at: &index, stringPtr: rowPtr),
             let absmag: Float = readNumber(at: &index, stringPtr: rowPtr) else { return nil }
         let spectralType = readString(at: &index, stringPtr: rowPtr)
